+++ src/events/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from fastapi_users_db_sqlalchemy import SQLAlchemyBaseUserTable
-from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text, Date
+from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text, Date, BIGINT
 
 from src.user_profile.models import user
 from src.club.models import club
@@ -23,7 +23,7 @@ event = Table(
 event_reg = Table(
     'event_reg', metadata,
     Column('id', Integer, primary_key=True),
-    Column('user_id', Integer, ForeignKey('user.id'), nullable=False),
+    Column('user_id', BIGINT, ForeignKey('user.id'), nullable=False),
     Column('event_id', Integer, ForeignKey('event.id'), nullable=False),
     # Column('confirm', Boolean, nullable=False, default=False),
     Column('reg_date', TIMESTAMP, nullable=False),
