+++ src/mentorship/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 
 from fastapi_users_db_sqlalchemy import SQLAlchemyBaseUserTable
-from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text
+from sqlalchemy import Table, Column, Integer, String, TIMESTAMP, ForeignKey, JSON, Boolean, Text, BIGINT
 
 from src.user_profile.models import user
 from src.club.models import club
@@ -13,7 +13,7 @@ mentorship = Table(
     Column('id', Integer, primary_key=True),
     Column('mentor_id', Integer, ForeignKey('user.id'), nullable=False),
     Column('mentee_id', Integer, ForeignKey('user.id'), nullable=False),
-    Column('club_id', Integer, ForeignKey('club.id'), nullable=False),
+    Column('club_id', BIGINT, ForeignKey('club.id'), nullable=False),
     # Column('start_date', TIMESTAMP, nullable=False),
     # Column('end_date', TIMESTAMP, nullable=False)
 )
